Rename `log4j-sdk` to `log4j-kit`
