+++ log4j-kit/src/main/java/org/apache/logging/log4j/kit/logger/package-info.java
@@ -16,7 +16,7 @@
  */
 @Export
 @Version("3.0.0")
-package org.apache.logging.log4j.sdk.logger;
+package org.apache.logging.log4j.kit.logger;
 
 import org.osgi.annotation.bundle.Export;
 import org.osgi.annotation.versioning.Version;
