+++ log4j-kit/src/test/java/org/apache/logging/log4j/kit/logger/AbstractLoggerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.sdk.logger;
+package org.apache.logging.log4j.kit.logger;
 
 import static org.assertj.core.api.Assertions.assertThat;
